+++ CHANGELOG.md
@@ -20,6 +20,12 @@
   - 🚨 [breaking change] Remove `backgroundColor` and `themeTooltip` props.
   - [feature] Add `coloring`, `padded`, and `maxWidth` props.
   - [feature] Children of Tooltip no longer need to be functions! If you provide a standard DOM element (e.g. `<button>`, `<div>`, `<span>`) or a Button component, things will just work. You can still use a function if your trigger is a custom component.
+- **Modal**
+  - 🚨 [breaking change] Remove `underlayStyle`, `underlayProps`, `themeUnderlay`, `themeModal`, and `themeButtonClose` props.
+  - 🚨 [breaking change] Rename `title` to `accessibleTitle`.
+  - [feature] Add `padded` prop.
+  - [fix] Modals taller than the viewport will now show padding above them: the browser will not longer automatically scroll down to the top of the modal.
+  - [fix] Update `react-aria-modal` to get some bug fixes and focus-management improvements.
 
 ## 0.2.1